Incorporate revew feedback; move Zip2Pull into test file.
