+++ transform.go
@@ -99,28 +99,8 @@ type Zipped[T1, T2 any] struct {
 
 // Zip returns a new Seq that yields the values of seq1 and seq2
 // simultaneously.
-func Zip2Pull[T1, T2 any](seq1 Seq[T1], seq2 Seq[T2]) Seq[Zipped[T1, T2]] {
-	return func(yield func(Zipped[T1, T2]) bool) {
-		p1, stop := Pull(seq1)
-		defer stop()
-		p2, stop := Pull(seq2)
-		defer stop()
-
-		for {
-			var val Zipped[T1, T2]
-			val.V1, val.OK1 = p1()
-			val.V2, val.OK2 = p2()
-			if (!val.OK1 && !val.OK2) || !yield(val) {
-				return
-			}
-		}
-	}
-}
-
-// Zip returns a new Seq that yields the values of seq1 and seq2
-// simultaneously.
 func Zip[T1, T2 any](seq1 Seq[T1], seq2 Seq[T2]) Seq[Zipped[T1, T2]] {
-	return func(body func(Zipped[T1, T2]) bool) {
+	return func(yield func(Zipped[T1, T2]) bool) {
 		p2, stop2 := Pull(seq2)
 		defer stop2()
 		done := false
@@ -129,7 +109,7 @@ func Zip[T1, T2 any](seq1 Seq[T1], seq2 Seq[T2]) Seq[Zipped[T1, T2]] {
 			var val Zipped[T1, T2]
 			val.V1, val.OK1 = v, true
 			val.V2, val.OK2 = p2()
-			if !body(val) {
+			if !yield(val) {
 				done = true
 				return false
 			}
@@ -145,7 +125,7 @@ func Zip[T1, T2 any](seq1 Seq[T1], seq2 Seq[T2]) Seq[Zipped[T1, T2]] {
 			var val Zipped[T1, T2]
 			val.V1, val.OK1 = v1, false
 			val.V2, val.OK2 = v2, true
-			if !body(val) {
+			if !yield(val) {
 				return
 			}
 		}
